tinder/requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.5
appnope==0.1.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
billiard==3.6.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.7
django-cleanup==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
geoip2==3.0.0
idna==2.10
inflection==0.3.1
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
lazy-object-proxy==1.6.0
maxminddb==2.0.3
multidict==5.1.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.10.0
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-assert-utils==0.2.2
pytest-common-subject==1.0.5
pytest-fixture-order==0.1.3
pytest-lambda==1.2.4
python-dateutil==2.8.1
python-env-utils==0.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
